Moved some methods about sending data to VR to the VRManager and added some comments
